To-Do List
In this project i've tried to make a to-do list with JavaScript. I made it but when i refresh the page it resets. Then i searched about localStorage and added to my project. Now when i refresh it doesn't resets. But there's a small issue with it. The sample "Task" task is reappearing with every refresh. Other things works well.